FORMS-17107: incorporated review comments
