+++ ui.frontend/src/utils.js
@@ -19,6 +19,7 @@ import HTTPAPILayer from "./HTTPAPILayer.js";
 import {customFunctions} from "./customFunctions.js";
 import {FunctionRuntime} from '@aemforms/af-core';
 import {loadXfa} from "./handleXfa";
+import RuleUtils from "./ruleUtils.js";
 
 /**
  * @module FormView
@@ -270,55 +271,6 @@ class Utils {
     }
 
     /**
-     * @deprecated Use `registerCustomFunctionsV2` instead.
-     * Registers custom functions from clientlibs.
-     * @param {string} formId - The form ID.
-     */
-    static async registerCustomFunctions(formId) {
-        const funcConfig = await HTTPAPILayer.getCustomFunctionConfig(formId);
-        console.debug("Fetched custom functions: " + JSON.stringify(funcConfig));
-        if (funcConfig && funcConfig.customFunction) {
-            const funcObj = funcConfig.customFunction.reduce((accumulator, func) => {
-                if (window[func.id]) {
-                    accumulator[func.id] = window[func.id];
-                }
-                return accumulator;
-            }, {});
-            FunctionRuntime.registerFunctions(funcObj);
-        }
-    }
-
-    /**
-     * Registers custom functions from clientlibs.
-     * @param {object} formJson - The Sling model exporter representation of the form
-     */
-    static async registerCustomFunctionsV2(formJson) {
-        let funcConfig;
-        let customFunctions = [];
-        const customFunctionsUrl = formJson.properties['fd:customFunctionsUrl'];
-        if (customFunctionsUrl) {
-            funcConfig = await HTTPAPILayer.getJson(customFunctionsUrl);
-        } else {
-            funcConfig = await HTTPAPILayer.getCustomFunctionConfig(formJson.id);
-        }
-        console.debug("Fetched custom functions: " + JSON.stringify(funcConfig));
-        
-        if (funcConfig && funcConfig.clientSideParsingEnabled) {
-            customFunctions = this.extractFunctionNames(formJson);
-        } else if (funcConfig && funcConfig.customFunction) {
-            customFunctions = funcConfig.customFunction;
-        }
-
-        const funcObj = customFunctions.reduce((accumulator, func) => {
-            if (window[func.id] && typeof window[func.id] === 'function') {
-                accumulator[func.id] = window[func.id];
-            }
-            return accumulator;
-        }, {});
-        FunctionRuntime.registerFunctions(funcObj);
-    }
-
-    /**
      * Sets up the Form Container.
      * @param {Function} createFormContainer - The function to create a form container.
      * @param {string} formContainerSelector - The CSS selector to identify the form container.
@@ -356,8 +308,8 @@ class Utils {
                     _formJson = await HTTPAPILayer.getFormDefinition(_path, _pageLang);
                 }
                 console.debug("fetched model json", _formJson);
-                await this.registerCustomFunctionsV2( _formJson);
-                await this.registerCustomFunctionsByUrl(customFunctionUrl);
+                await RuleUtils.registerCustomFunctionsV2( _formJson);
+                await RuleUtils.registerCustomFunctionsByUrl(customFunctionUrl);
                 const urlSearchParams = new URLSearchParams(window.location.search);
                 const params = Object.fromEntries(urlSearchParams.entries());
                 let _prefillData = {};
@@ -387,94 +339,6 @@ class Utils {
                 Utils.initializeAllFields(formContainer);
                 const event = new CustomEvent(Constants.FORM_CONTAINER_INITIALISED, { "detail": formContainer });
                 document.dispatchEvent(event);
-            }
-        }
-    }
-
-    
-    /**
-     * Extract custom functions from form json and return them in the required format.
-     * @param {object} formJson - The Sling model exporter representation of the form
-     * @returns {Array<{id: string}>} - Array of objects with function names as ids
-     */
-    static extractFunctionNames = (formJson) => {
-        const functionNames = new Set();
-        const exclusionList = ['dispatchEvent'];
-        
-        // Helper function to extract function names from event strings
-        const extractFromEventString = (eventString) => {
-            // Match patterns like functionName() with optional parameters
-            const functionPattern = /(\w+)\s*\(/g;
-            let match;
-            
-            while ((match = functionPattern.exec(eventString)) !== null) {
-                // Add the function name to our set
-                if (!exclusionList.includes(match[1])) {
-                    functionNames.add(match[1]);
-                }
-            }
-        };
-        
-        // Process events at the form level
-        if (formJson.events) {
-            Object.values(formJson.events).forEach(eventArray => {
-                eventArray.forEach(eventString => {
-                    extractFromEventString(eventString);
-                });
-            });
-        }
-        
-        // Recursively process events in form items
-        const processItems = (items) => {
-            if (!items) return;
-            
-            Object.values(items).forEach(item => {
-                if (item.events) {
-                    Object.values(item.events).forEach(eventArray => {
-                        eventArray.forEach(eventString => {
-                            extractFromEventString(eventString);
-                        });
-                    });
-                }
-                
-                // Process nested items if they exist
-                if (item[':items']) {
-                    processItems(item[':items']);
-                }
-            });
-        };
-        
-        if (formJson[':items']) {
-            processItems(formJson[':items']);
-        }
-
-        console.debug("Extracted custom function names from form JSON:", functionNames);
-
-        // Convert the array of function names to the required format
-        return Array.from(functionNames).map(functionName => ({
-            id: functionName
-        }));
-    };
-
-    static async registerCustomFunctionsByUrl(url) {
-        try {
-            if (url != null && url.trim().length > 0) {
-                // webpack ignore is added because webpack was converting this to a static import upon bundling resulting in error.
-                //This Url should whitelist the AEM author/publish domain in the Cross Origin Resource Sharing (CORS) configuration.
-                const customFunctionModule = await import(/*webpackIgnore: true*/ url);
-                const keys = Object.keys(customFunctionModule);
-                const functions = [];
-                for (const name of keys) {
-                    const funcDef = customFunctionModule[name];
-                    if (typeof funcDef === 'function') {
-                        functions[name] = funcDef;
-                    }
-                }
-                FunctionRuntime.registerFunctions(functions);
-            }
-        } catch (e) {
-            if(window.console){
-                console.error("error in loading custom functions from url "+url+" with message "+e.message);
             }
         }
     }
